refactor: foldersPicker
